+++ task3.py
@@ -1,11 +1,11 @@
-#! python3 
+#! python3
 
-# Have the user enter a number. 
-# Use an if...elif statement to determine if the number is 
-# a) larger than 1000 
-# b) larger than 100 
-# c) larger than 10 
-# d) larger than 0 
+# Have the user enter a number.
+# Use an if...elif statement to determine if the number is
+# a) larger than 1000
+# b) larger than 100
+# c) larger than 10
+# d) larger than 0
 # Output must match one of the valid output statements listed
 # (2 points)
 
@@ -14,12 +14,12 @@
 
 # Output is a single number that represents a range of numbers:
 # "3" : The number is equal to 1000 or is larger than 1000
-# "2" : The number is 100 or a number up to 1000 
-# "1" : The number is 10 or a number up to 100 
-# "0" : The number is 0 or a number up to 100 
+# "2" : The number is 100 or a number up to 1000
+# "1" : The number is 10 or a number up to 100
+# "0" : The number is 0 or a number up to 100
 
 inputnumber = input("Enter a number: ")
-x = int(inputnumber)
+x = float(inputnumber)
 a = 1000
 b = 100
 c = 10
@@ -44,4 +44,3 @@ elif x > d:
     print("0")
 elif x == d:
     print("0")
-
